Abstract draftjs auto embed parsing
